+++ api/src/main/java/net/jqwik/api/constraints/NotBlank.java
@@ -1,0 +1,22 @@
+package net.jqwik.api.constraints;
+
+import java.lang.annotation.*;
+
+import org.apiguardian.api.*;
+
+import static org.apiguardian.api.API.Status.*;
+
+/**
+ * Constrain a string to never be blank, i.e. not empty and not just whitespace.
+ *
+ * Applies to String parameters which are also annotated with {@code @ForAll}.
+ *
+ * @see net.jqwik.api.ForAll
+ * @see NotEmpty
+ */
+@Target({ ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE })
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@API(status = EXPERIMENTAL, since = "1.5.0")
+public @interface NotBlank {
+}